+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new hidden UI for administrating canvases which lets you re-ingest a thumbnail if it had previously failed.
 - Local tool (for now) to build pre-release docker tags, without creating a latest tag to be manually pushed to Dockerhub.
 - Added new CLI tool for getting all of the Madoc templates that can be added into a theme.
+- Added basic public user profiles
+- Added option to show email on public user profiles (logged in only)
+- Added option for semi-public profiles, requiring authentication
 
 ### Fixes
 - Fixed bug where canvas ID list may be only a single element